pyproject.toml
[project]
name = "business-brain"
version = "0.1.0"
description = "The Business Brain — an autonomous AI agent that acts as a proactive business analyst"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy[asyncio]>=2.0",
    "psycopg[binary]>=3.1",
    "pgvector>=0.3",
    "langchain>=0.3",
    "langgraph>=0.2",
    "watchdog>=4.0",
    "pandas>=2.1",
    "scipy>=1.11",
    "numpy>=1.26",
    "pydantic-settings>=2.1",
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "httpx>=0.27",
    "google-genai>=1.0",
    "python-multipart>=0.0.7",
    "openpyxl>=3.1",
    "PyPDF2>=3.0",
    "bcrypt>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/business_brain"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
